@@ -15,11 +15,13 @@
         "forum": "https://discourse.laminas.dev/"
     },
     "require": {
-        "php": "^5.6 || ^7.0 || ^8.0"
+        "php": "^7.3 || ^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^5.7 || ^6.5 || ^7.5 || ^8.1 || ^9.3",
-        "squizlabs/php_codesniffer": "^3.5"
+        "psalm/plugin-phpunit": "^0.15.1",
+        "squizlabs/php_codesniffer": "^3.5",
+        "vimeo/psalm": "^4.6"
     },
     "autoload": {
         "files": [
@@ -52,6 +54,7 @@
     "scripts": {
         "cs-check": "phpcs",
         "cs-fix": "phpcbf",
+        "static-analysis": "psalm --shepherd --stats",
         "test": "phpunit --colors=always",
         "test-coverage": "phpunit --colors=always --coverage-clover clover.xml"
     }
