@@ -24,7 +24,7 @@ use Grav\Common\Page\Pages;
 use Grav\Common\Page\Collection;
 use Grav\Common\Security;
 use Grav\Common\User\Interfaces\UserCollectionInterface;
-use Grav\Common\User\User;
+use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Utils;
 use Grav\Framework\Psr7\Response;
 use Grav\Framework\RequestHandler\Exception\RequestException;
@@ -36,7 +36,6 @@ use RocketTheme\Toolbox\Event\Event;
 use RocketTheme\Toolbox\File\File;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use Twig\Loader\FilesystemLoader;
-
 
 /**
  * Class AdminController
@@ -51,6 +50,7 @@ class AdminController extends AdminBaseController
      * @param string|null $task
      * @param string|null $route
      * @param array|null $post
+     * @return void
      */
     public function initialize(Grav $grav = null, $view = null, $task = null, $route = null, $post = null)
     {
@@ -76,10 +76,13 @@ class AdminController extends AdminBaseController
     /**
      * Keep alive
      *
+     * Route: POST /task:keepAlive (AJAX call)
+     *
      * @return void
      */
     protected function taskKeepAlive(): void
     {
+        // This task is available for all admin users.
         $response = new Response(200);
 
         $this->close($response);
@@ -88,11 +91,18 @@ class AdminController extends AdminBaseController
     /**
      * Clear the cache.
      *
+     * Route: GET /cache.json/task:clearCache (AJAX call)
+     *
      * @return bool True if the action was performed.
      */
     protected function taskClearCache()
     {
-        if (!$this->authorizeTask('clear cache', ['admin.cache', 'admin.super', 'admin.maintenance'])) {
+        if (!$this->authorizeTask('clear cache', ['admin.cache', 'admin.maintenance', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -132,6 +142,10 @@ class AdminController extends AdminBaseController
     /**
      * Handles form and saves the input data if its valid.
      *
+     * Route: POST /pages?task:save
+     * Route: POST /user?task:save
+     * Route: POST /*?task:save
+     *
      * @return bool True if the action was performed.
      */
     public function taskSave()
@@ -144,18 +158,20 @@ class AdminController extends AdminBaseController
 
         switch ($this->view) {
             case 'pages':
-                return $this->taskSavePage();
+                // Not used if Flex-Objects plugin handles pages.
+                return $this->savePage();
             case 'user':
-                return $this->taskSaveUser();
+                // Not used if Flex-Objects plugin handles users.
+                return $this->saveUser();
             default:
-                return $this->taskSaveDefault();
+                return $this->saveDefault();
         }
     }
 
     /**
      * @return bool
      */
-    protected function taskSaveDefault()
+    protected function saveDefault()
     {
         // Handle standard data types.
         $type = $this->getDataType();
@@ -195,53 +211,43 @@ class AdminController extends AdminBaseController
         return true;
     }
 
-     // LOGIN & USER TASKS
-
-    /**
-     * Handle login.
-     *
-     * @return bool True if the action was performed.
-     */
-    protected function taskLogin()
-    {
-        $this->admin->authenticate($this->data, $this->post);
-
-        return true;
-    }
-
-    /**
-     * @return bool True if the action was performed.
-     */
-    protected function taskTwofa()
-    {
-        $this->admin->twoFa($this->data, $this->post);
-
-        return true;
-    }
+    // USER TASKS
 
     /**
      * Handle logout.
+     *
+     * Route: GET /task:logout
+     * Route: POST ?task=logout
      *
      * @return bool True if the action was performed.
      */
     protected function taskLogout()
     {
-        $this->admin->logout($this->data, $this->post);
+        if (!$this->authorizeTask('logout', ['admin.login', 'admin.super'])) {
+            return false;
+        }
 
-        return true;
+        $this->admin->logout($this->data, $this->post);
     }
 
     /**
+     * Route: POST /ajax.json/task:regenerate2FASecret (AJAX call)
+     *
      * @return bool
      */
     public function taskRegenerate2FASecret()
     {
-        if (!$this->authorizeTask('regenerate 2FA Secret', ['admin.login'])) {
+        if (!$this->authorizeTask('regenerate 2FA Secret', ['admin.login', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
         try {
-            /** @var User $user */
+            /** @var UserInterface $user */
             $user = $this->grav['user'];
 
             /** @var TwoFactorAuth $twoFa */
@@ -279,179 +285,15 @@ class AdminController extends AdminBaseController
     }
 
     /**
-     * Handle the reset password action.
-     *
-     * @return bool True if the action was performed.
-     */
-    public function taskReset()
-    {
-        $data = $this->data;
-
-        if (isset($data['password'])) {
-            /** @var UserCollectionInterface $users */
-            $users = $this->grav['accounts'];
-
-            $username = isset($data['username']) ? strip_tags(strtolower($data['username'])) : null;
-            $user     = $username ? $users->load($username) : null;
-            $password = $data['password'] ?? null;
-            $token    = $data['token'] ?? null;
-
-            if ($user && $user->exists() && !empty($user->get('reset'))) {
-                list($good_token, $expire) = explode('::', $user->get('reset'));
-
-                if ($good_token === $token) {
-                    if (time() > $expire) {
-                        $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.RESET_LINK_EXPIRED'), 'error');
-                        $this->setRedirect('/forgot');
-
-                        return true;
-                    }
-
-                    $user->undef('hashed_password');
-                    $user->undef('reset');
-                    $user->set('password',  $password);
-                    $user->save();
-
-                    $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.RESET_PASSWORD_RESET'), 'info');
-                    $this->setRedirect('/');
-
-                    return true;
-                }
-            }
-
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.RESET_INVALID_LINK'), 'error');
-            $this->setRedirect('/forgot');
-
-            return true;
-
-        }
-
-        $user  = $this->grav['uri']->param('user');
-        $token = $this->grav['uri']->param('token');
-
-        if (empty($user) || empty($token)) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.RESET_INVALID_LINK'), 'error');
-            $this->setRedirect('/forgot');
-
-            return true;
-        }
-
-        $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.RESET_NEW_PASSWORD'), 'info');
-
-        $this->admin->forgot = ['username' => $user, 'token' => $token];
-
-        return true;
-    }
-
-    /**
-     * Handle the email password recovery procedure.
-     *
-     * @return bool True if the action was performed.
-     */
-    protected function taskForgot()
-    {
-        $param_sep = $this->grav['config']->get('system.param_sep', ':');
-        $post      = $this->post;
-        $data      = $this->data;
-        $login     = $this->grav['login'];
-
-        /** @var UserCollectionInterface $users */
-        $users = $this->grav['accounts'];
-
-        $username = isset($data['username']) ? strip_tags(strtolower($data['username'])) : '';
-        $user     = !empty($username) ? $users->load($username) : null;
-
-        if (!isset($this->grav['Email'])) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_EMAIL_NOT_CONFIGURED'), 'error');
-            $this->setRedirect($post['redirect']);
-
-            return true;
-        }
-
-        if (!$user || !$user->exists()) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_INSTRUCTIONS_SENT_VIA_EMAIL'),
-                'info');
-            $this->setRedirect($post['redirect']);
-
-            return true;
-        }
-
-        if (empty($user->email)) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_INSTRUCTIONS_SENT_VIA_EMAIL'),
-                'info');
-            $this->setRedirect($post['redirect']);
-
-            return true;
-        }
-
-        $count = $this->grav['config']->get('plugins.login.max_pw_resets_count', 0);
-        $interval =$this->grav['config']->get('plugins.login.max_pw_resets_interval', 2);
-
-        if ($login->isUserRateLimited($user, 'pw_resets', $count, $interval)) {
-            $this->admin->setMessage($this->admin::translate(['PLUGIN_LOGIN.FORGOT_CANNOT_RESET_IT_IS_BLOCKED', $user->email, $interval]), 'error');
-            $this->setRedirect($post['redirect']);
-
-            return true;
-        }
-
-        $token  = md5(uniqid(mt_rand(), true));
-        $expire = time() + 604800; // next week
-
-        $user->set('reset', $token . '::' . $expire);
-        $user->save();
-
-        $author     = $this->grav['config']->get('site.author.name', '');
-        $fullname   = $user->fullname ?: $username;
-        $reset_link = rtrim($this->grav['uri']->rootUrl(true), '/') . '/' . trim($this->admin->base,
-                '/') . '/reset/task' . $param_sep . 'reset/user' . $param_sep . $username . '/token' . $param_sep . $token . '/admin-nonce' . $param_sep . Utils::getNonce('admin-form');
-
-        $sitename = $this->grav['config']->get('site.title', 'Website');
-        $from     = $this->grav['config']->get('plugins.email.from');
-
-        if (empty($from)) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_EMAIL_NOT_CONFIGURED'), 'error');
-            $this->setRedirect($post['redirect']);
-
-            return true;
-        }
-
-        $to = $user->email;
-
-        $subject = $this->admin::translate(['PLUGIN_ADMIN.FORGOT_EMAIL_SUBJECT', $sitename]);
-        $content = $this->admin::translate([
-            'PLUGIN_ADMIN.FORGOT_EMAIL_BODY',
-            $fullname,
-            $reset_link,
-            $author,
-            $sitename
-        ]);
-
-        $body = $this->grav['twig']->processTemplate('email/base.html.twig', ['content' => $content]);
-
-        $message = $this->grav['Email']->message($subject, $body, 'text/html')->setFrom($from)->setTo($to);
-
-        $sent = $this->grav['Email']->send($message);
-
-        if ($sent < 1) {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_FAILED_TO_EMAIL'), 'error');
-        } else {
-            $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.FORGOT_INSTRUCTIONS_SENT_VIA_EMAIL'),
-                'info');
-        }
-
-        $this->setRedirect('/');
-
-        return true;
-    }
-
-    /**
      * Save user account.
      *
      * Called by more general save task.
      *
+     * @note Not used if Flex-Objects plugin handles users.
+     *
      * @return bool
      */
-    protected function taskSaveUser()
+    protected function saveUser()
     {
         /** @var UserCollectionInterface $users */
         $users = $this->grav['accounts'];
@@ -514,12 +356,19 @@ class AdminController extends AdminBaseController
     /**
      * Get Notifications
      *
-     * @return void
+     * Route: POST /task:getNotifications (AJAX call)
+     *
+     * @return bool
      */
-    protected function taskGetNotifications(): void
+    protected function taskGetNotifications()
     {
         if (!$this->authorizeTask('dashboard', ['admin.login', 'admin.super'])) {
-            $this->sendJsonResponse(['status' => 'error', 'message' => 'unauthorized']);
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
+            return false;
         }
 
         // do we need to force a reload
@@ -557,11 +406,18 @@ class AdminController extends AdminBaseController
     /**
      * Hide notifications.
      *
+     * Route: POST /notifications.json/task:hideNotification/notification_id:ID (AJAX call)
+     *
      * @return bool True if the action was performed.
      */
     protected function taskHideNotification()
     {
-        if (!$this->authorizeTask('hide notification', ['admin.login'])) {
+        if (!$this->authorizeTask('hide notification', ['admin.login', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -593,12 +449,19 @@ class AdminController extends AdminBaseController
     /**
      * Get Newsfeeds
      *
-     * @return void
+     * Route: POST /ajax.json/task:getNewsFeed (AJAX call)
+     *
+     * @return bool
      */
-    protected function taskGetNewsFeed(): void
+    protected function taskGetNewsFeed()
     {
         if (!$this->authorizeTask('dashboard', ['admin.login', 'admin.super'])) {
-            $this->sendJsonResponse(['status' => 'error', 'message' => 'unauthorized']);
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
+            return false;
         }
 
         $refresh = $this->data['refresh'] === 'true' ? true : false;
@@ -627,15 +490,22 @@ class AdminController extends AdminBaseController
     /**
      * Handle the backup action
      *
+     * Route: GET /backup.json/id:BACKUP_ID/task:backup (AJAX call)
+     *
      * @return bool True if the action was performed.
      */
     protected function taskBackup()
     {
-        $param_sep = $this->grav['config']->get('system.param_sep', ':');
         if (!$this->authorizeTask('backup', ['admin.maintenance', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
+        $param_sep = $this->grav['config']->get('system.param_sep', ':');
         $download = $this->grav['uri']->param('download');
 
         try {
@@ -685,11 +555,18 @@ class AdminController extends AdminBaseController
     /**
      * Handle delete backup action
      *
+     * Route: GET /backup.json/backup:BACKUP_FILE/task:backupDelete (AJAX call)
+     *
      * @return bool
      */
     protected function taskBackupDelete()
     {
         if (!$this->authorizeTask('backup', ['admin.maintenance', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -727,17 +604,17 @@ class AdminController extends AdminBaseController
     /**
      * Enable a plugin.
      *
-     * Route: /plugins
+     * Route: GET /plugins/SLUG/task:enable
      *
      * @return bool True if the action was performed.
      */
     public function taskEnable()
     {
-        if (!$this->authorizeTask('enable plugin', ['admin.plugins', 'admin.super'])) {
+        if ($this->view !== 'plugins') {
             return false;
         }
 
-        if ($this->view !== 'plugins') {
+        if (!$this->authorizeTask('enable plugin', ['admin.plugins', 'admin.super'])) {
             return false;
         }
 
@@ -757,17 +634,17 @@ class AdminController extends AdminBaseController
     /**
      * Disable a plugin.
      *
-     * Route: /plugins
+     * Route: GET /plugins/SLUG/task:disable
      *
      * @return bool True if the action was performed.
      */
     public function taskDisable()
     {
-        if (!$this->authorizeTask('disable plugin', ['admin.plugins', 'admin.super'])) {
+        if ($this->view !== 'plugins') {
             return false;
         }
 
-        if ($this->view !== 'plugins') {
+        if (!$this->authorizeTask('disable plugin', ['admin.plugins', 'admin.super'])) {
             return false;
         }
 
@@ -787,17 +664,17 @@ class AdminController extends AdminBaseController
     /**
      * Set the default theme.
      *
-     * Route: /themes
+     * Route: GET /themes/SLUG/task:activate
      *
      * @return bool True if the action was performed.
      */
     public function taskActivate()
     {
-        if (!$this->authorizeTask('activate theme', ['admin.themes', 'admin.super'])) {
+        if ($this->view !== 'themes') {
             return false;
         }
 
-        if ($this->view !== 'themes') {
+        if (!$this->authorizeTask('activate theme', ['admin.themes', 'admin.super'])) {
             return false;
         }
 
@@ -831,11 +708,18 @@ class AdminController extends AdminBaseController
     /**
      * Handles updating Grav
      *
+     * Route: GET /update.json/task:updategrav (AJAX call)
+     *
      * @return bool False if user has no permissions.
      */
     public function taskUpdategrav()
     {
         if (!$this->authorizeTask('install grav', ['admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -860,23 +744,21 @@ class AdminController extends AdminBaseController
         }
 
         $this->sendJsonResponse($json_response);
-
-        return true;
     }
 
     /**
      * Handles uninstalling plugins and themes
      *
-     * Route: /plugins
-     * Route: /themes
-     *
-     * @deprecated
-     *
      * @return bool True if the action was performed
+     * @deprecated Not being used anymore
      */
     public function taskUninstall()
     {
-        $type = $this->view === 'plugins' ? 'plugins' : 'themes';
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
+
         if (!$this->authorizeTask('uninstall ' . $type, ['admin.' . $type, 'admin.super'])) {
             return false;
         }
@@ -899,11 +781,18 @@ class AdminController extends AdminBaseController
     /**
      * Toggle the gpm.releases setting
      *
+     * Route: POST /ajax.json/task:gpmRelease (AJAX call)
+     *
      * @return bool
      */
     protected function taskGpmRelease()
     {
-        if (!$this->authorizeTask('configuration', ['admin.configuration.system', 'admin.super'])) {
+        if (!$this->authorizeTask('configuration', ['admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -942,11 +831,22 @@ class AdminController extends AdminBaseController
     /**
      * Get update status from GPM
      *
+     * Request: POST /update.json/task:getUpdates (AJAX call)
+     *
      * @return bool
      */
     protected function taskGetUpdates()
     {
-        if (!$this->authorizeTask('dashboard', ['admin.login', 'admin.super'])) {
+        if ($this->view !== 'update') {
+            return false;
+        }
+
+        if (!$this->authorizeTask('dashboard', ['admin.plugins', 'admin.themes', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -1010,12 +910,29 @@ class AdminController extends AdminBaseController
     }
 
     /**
-     * Handle getting a new package dependencies needed to be installed
+     * Handle getting a new package dependencies needed to be installed.
+     *
+     * Route: /plugins.json/task:getPackagesDependencies (AJAX call)
+     * Route: /themes.json/task:getPackagesDependencies (AJAX call)
      *
      * @return bool
      */
     protected function taskGetPackagesDependencies()
     {
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
+
+        if (!$this->authorizeTask('get package dependencies', ['admin.' . $type, 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
+            return false;
+        }
+
         $data     = $this->post;
         $packages = isset($data['packages']) ? explode(',', $data['packages']) : '';
         $packages = (array)$packages;
@@ -1039,17 +956,19 @@ class AdminController extends AdminBaseController
     }
 
     /**
+     * Route: /plugins.json/task:installDependenciesOfPackages (AJAX call)
+     * Route: /themes.json/task:installDependenciesOfPackages (AJAX call)
+     *
      * @return bool
      */
     protected function taskInstallDependenciesOfPackages()
     {
-        $data     = $this->post;
-        $packages = isset($data['packages']) ? explode(',', $data['packages']) : '';
-        $packages = (array)$packages;
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
 
-        $type = $data['type'] ?? '';
-
-        if (!$this->authorizeTask('install ' . $type, ['admin.' . $type, 'admin.super'])) {
+        if (!$this->authorizeTask('install dependencies', ['admin.' . $type, 'admin.super'])) {
             $this->admin->json_response = [
                 'status'  => 'error',
                 'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
@@ -1057,6 +976,10 @@ class AdminController extends AdminBaseController
 
             return false;
         }
+
+        $data = $this->post;
+        $packages = isset($data['packages']) ? explode(',', $data['packages']) : '';
+        $packages = (array)$packages;
 
         try {
             $dependencies = $this->admin->getDependenciesNeededToInstall($packages);
@@ -1085,14 +1008,18 @@ class AdminController extends AdminBaseController
     }
 
     /**
+     * Route: /plugins.json/task:installPackage (AJAX call)
+     * Route: /themes.json/task:installPackage (AJAX call)
+     *
      * @param bool $reinstall
      * @return bool
      */
     protected function taskInstallPackage($reinstall = false)
     {
-        $data    = $this->post;
-        $package = $data['package'] ?? '';
-        $type    = $data['type'] ?? '';
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
 
         if (!$this->authorizeTask('install ' . $type, ['admin.' . $type, 'admin.super'])) {
             $this->admin->json_response = [
@@ -1103,6 +1030,8 @@ class AdminController extends AdminBaseController
             return false;
         }
 
+        $data = $this->post;
+        $package = $data['package'] ?? '';
         try {
             $result = Gpm::install($package, ['theme' => $type === 'theme']);
         } catch (\Exception $e) {
@@ -1138,14 +1067,17 @@ class AdminController extends AdminBaseController
     /**
      * Handle removing a package
      *
-     * @return void
+     * Route: /plugins.json/task:removePackage (AJAX call)
+     * Route: /themes.json/task:removePackage (AJAX call)
+     *
+     * @return bool
      */
-    protected function taskRemovePackage(): void
+    protected function taskRemovePackage(): bool
     {
-        $data    = $this->post;
-        $package = $data['package'] ?? '';
-        $type    = $data['type'] ?? '';
-        $result = false;
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
 
         if (!$this->authorizeTask('uninstall ' . $type, ['admin.' . $type, 'admin.super'])) {
             $json_response = [
@@ -1155,6 +1087,10 @@ class AdminController extends AdminBaseController
 
             $this->sendJsonResponse($json_response, 403);
         }
+
+        $data    = $this->post;
+        $package = $data['package'] ?? '';
+        $result = false;
 
         //check if there are packages that have this as a dependency. Abort and show which ones
         $dependent_packages = $this->admin->getPackagesThatDependOnPackage($package);
@@ -1207,16 +1143,17 @@ class AdminController extends AdminBaseController
     /**
      * Handle reinstalling a package
      *
-     * @return void
+     * Route: /plugins.json/task:reinstallPackage (AJAX call)
+     * Route: /themes.json/task:reinstallPackage (AJAX call)
+     *
+     * @return bool
      */
     protected function taskReinstallPackage()
     {
-        $data = $this->post;
-
-        $slug            = $data['slug'] ?? '';
-        $type            = $data['type'] ?? '';
-        $package_name    = $data['package_name'] ?? '';
-        $current_version = $data['current_version'] ?? '';
+        $type = $this->view;
+        if ($type !== 'plugins' && $type !== 'themes') {
+            return false;
+        }
 
         if (!$this->authorizeTask('install ' . $type, ['admin.' . $type, 'admin.super'])) {
             $json_response = [
@@ -1226,6 +1163,11 @@ class AdminController extends AdminBaseController
 
             $this->sendJsonResponse($json_response, 403);
         }
+
+        $data = $this->post;
+        $slug            = $data['slug'] ?? '';
+        $package_name    = $data['package_name'] ?? '';
+        $current_version = $data['current_version'] ?? '';
 
         $url = "https://getgrav.org/download/{$type}s/$slug/$current_version";
 
@@ -1243,10 +1185,14 @@ class AdminController extends AdminBaseController
                 'message' => $this->admin::translate('PLUGIN_ADMIN.REINSTALLATION_FAILED')
             ];
         }
+
+        return true;
     }
 
     /**
      * Handle direct install.
+     *
+     * Request: POST /tools/direct-install?task=directInstall
      *
      * @return bool
      */
@@ -1313,11 +1259,19 @@ class AdminController extends AdminBaseController
     /**
      * Handles creating an empty page folder (without markdown file)
      *
+     * Route: /pages
+     *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool True if the action was performed.
      */
     public function taskSaveNewFolder()
     {
-        if (!$this->authorizeTask('save', $this->dataPermissions())) {
+        if ($this->view !== 'pages') {
+            return false;
+        }
+
+        if (!$this->authorizeTask('new folder', ['admin.pages', 'admin.super'])) {
             return false;
         }
 
@@ -1356,9 +1310,11 @@ class AdminController extends AdminBaseController
     }
 
     /**
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool
      */
-    protected function taskSavePage()
+    protected function savePage()
     {
         $reorder = true;
 
@@ -1498,17 +1454,18 @@ class AdminController extends AdminBaseController
      *
      * Route: /pages
      *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool True if the action was performed.
      * @throws \RuntimeException
      */
     protected function taskCopy()
     {
-        if (!$this->authorizeTask('copy page', ['admin.pages', 'admin.super'])) {
+        if ($this->view !== 'pages') {
             return false;
         }
 
-        // Only applies to pages.
-        if ($this->view !== 'pages') {
+        if (!$this->authorizeTask('copy page', ['admin.pages', 'admin.super'])) {
             return false;
         }
 
@@ -1587,16 +1544,17 @@ class AdminController extends AdminBaseController
      *
      * Route: /pages
      *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool True if the action was performed.
      */
     protected function taskReorder()
     {
-        if (!$this->authorizeTask('reorder pages', ['admin.pages', 'admin.super'])) {
+        if ($this->view !== 'pages') {
             return false;
         }
 
-        // Only applies to pages.
-        if ($this->view !== 'pages') {
+        if (!$this->authorizeTask('reorder pages', ['admin.pages', 'admin.super'])) {
             return false;
         }
 
@@ -1610,17 +1568,18 @@ class AdminController extends AdminBaseController
      *
      * Route: /pages
      *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool True if the action was performed.
      * @throws \RuntimeException
      */
     protected function taskDelete()
     {
-        if (!$this->authorizeTask('delete page', ['admin.pages', 'admin.super'])) {
+        if ($this->view !== 'pages') {
             return false;
         }
 
-        // Only applies to pages.
-        if ($this->view !== 'pages') {
+        if (!$this->authorizeTask('delete page', ['admin.pages', 'admin.super'])) {
             return false;
         }
 
@@ -1656,10 +1615,16 @@ class AdminController extends AdminBaseController
      *
      * Route: /pages
      *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool
      */
     protected function taskSwitchlanguage()
     {
+        if ($this->view !== 'pages') {
+            return false;
+        }
+
         if (!$this->authorizeTask('switch language', ['admin.pages', 'admin.super'])) {
             return false;
         }
@@ -1689,11 +1654,19 @@ class AdminController extends AdminBaseController
     /**
      * Save the current page in a different language. Automatically switches to that language.
      *
+     * Route: /pages
+     *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool True if the action was performed.
      */
     protected function taskSaveas()
     {
-        if (!$this->authorizeTask('save', $this->dataPermissions())) {
+        if ($this->view !== 'pages') {
+            return false;
+        }
+
+        if (!$this->authorizeTask('save as', ['admin.pages', 'admin.super'])) {
             return false;
         }
 
@@ -1749,6 +1722,8 @@ class AdminController extends AdminBaseController
     /**
      * Continue to the new page.
      *
+     * @note Not used if Flex-Objects plugin handles pages, users and user groups.
+     *
      * @return bool True if the action was performed.
      */
     public function taskContinue()
@@ -1758,12 +1733,6 @@ class AdminController extends AdminBaseController
         if ($this->view === 'users') {
             $username = strip_tags(strtolower($data['username']));
             $this->setRedirect("{$this->view}/{$username}");
-
-            return true;
-        }
-
-        if ($this->view === 'groups') {
-            $this->setRedirect("{$this->view}/{$data['groupname']}");
 
             return true;
         }
@@ -1850,11 +1819,24 @@ class AdminController extends AdminBaseController
     }
 
     /**
+     * Route: /ajax.json
+     *
+     * @note Not used if Flex-Objects plugin handles pages.
+     *
      * @return bool
      */
     protected function taskGetChildTypes()
     {
+        if ($this->view !== 'ajax') {
+            return false;
+        }
+
         if (!$this->authorizeTask('get childtypes', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -1893,12 +1875,23 @@ class AdminController extends AdminBaseController
     /**
      * Handles filtering the page by modular/visible/routable in the pages list.
      *
-     * @return void
+     * @note Used only in legacy pages.
+     *
+     * @return bool
      */
     protected function taskFilterPages()
     {
+        if ($this->view !== 'pages-filter') {
+            return false;
+        }
+
         if (!$this->authorizeTask('filter pages', ['admin.pages', 'admin.super'])) {
-            return;
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
+            return false;
         }
 
         $data = $this->post;
@@ -2013,17 +2006,26 @@ class AdminController extends AdminBaseController
             'results' => $results
         ];
         $this->admin->collection = $collection;
+
+        return true;
     }
 
 
     /**
      * Process the page Markdown
      *
+     * Preview task in the builtin editor.
+     *
      * @return bool True if the action was performed.
      */
     protected function taskProcessMarkdown()
     {
         if (!$this->authorizeTask('process markdown', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -2068,13 +2070,22 @@ class AdminController extends AdminBaseController
     /**
      * Determines the file types allowed to be uploaded
      *
-     * Used by pagemedia field.
+     * Used by pagemedia field. Works only within legacy pages.
      *
      * @return bool True if the action was performed.
      */
     protected function taskListmedia()
     {
+        if ($this->view !== 'media') {
+            return false;
+        }
+
         if (!$this->authorizeTask('list media', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -2121,13 +2132,22 @@ class AdminController extends AdminBaseController
     /**
      * Handles adding a media file to a page.
      *
-     * Used by pagemedia field.
+     * Used by pagemedia field. Works only within legacy pages.
      *
      * @return bool True if the action was performed.
      */
     protected function taskAddmedia()
     {
+        if ($this->view !== 'media') {
+            return false;
+        }
+
         if (!$this->authorizeTask('add media', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -2286,12 +2306,18 @@ class AdminController extends AdminBaseController
     }
 
     /**
+     * Request: POST .json/task:compileScss (AJAX call)
+     *
      * @return bool
      */
     protected function taskCompileScss()
     {
+        if (!$this->authorizeTask('compile scss', ['admin.plugins', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
 
-        if (!$this->authorizeTask('compile scss', ['admin.super'])) {
             return false;
         }
 
@@ -2316,16 +2342,22 @@ class AdminController extends AdminBaseController
             ]
         ];
 
-        echo json_encode($json_response);
-        exit;
+        $this->sendJsonResponse($json_response);
     }
 
     /**
+     * Request: POST .json/task:exportScss (AJAX call)
+     *
      * @return bool
      */
     protected function taskExportScss()
     {
-        if (!$this->authorizeTask('compile scss', ['admin.pages', 'admin.super'])) {
+        if (!$this->authorizeTask('export scss', ['admin.plugins', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -2344,8 +2376,7 @@ class AdminController extends AdminBaseController
             ]
         ];
 
-        echo json_encode($json_response);
-        exit;
+        $this->sendJsonResponse($json_response);
     }
 
     /**
@@ -2357,7 +2388,16 @@ class AdminController extends AdminBaseController
      */
     protected function taskDelmedia()
     {
+        if ($this->view !== 'media') {
+            return false;
+        }
+
         if (!$this->authorizeTask('delete media', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
             return false;
         }
 
@@ -2639,10 +2679,6 @@ class AdminController extends AdminBaseController
      */
     public function getMedia(PageInterface $page = null)
     {
-        if ($this->view !== 'media') {
-            return null;
-        }
-
         $page = $page ?? $this->admin->page($this->route);
         if (!$page) {
             return null;
@@ -2937,10 +2973,21 @@ class AdminController extends AdminBaseController
     }
 
     /**
-     * @return ResponseInterface
+     * Used in 3rd party editors (e.g. next-gen).
+     *
+     * @return ResponseInterface|false
      */
-    protected function taskConvertUrls(): ResponseInterface
+    protected function taskConvertUrls()
     {
+        if (!$this->authorizeTask('access page', ['admin.pages', 'admin.super'])) {
+            $this->admin->json_response = [
+                'status'  => 'error',
+                'message' => $this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK')
+            ];
+
+            return false;
+        }
+
         $request = $this->getRequest();
         $data = $request->getParsedBody();
         $converted_links = [];
