@@ -2,8 +2,10 @@
 
 namespace Grav\Plugin\Admin;
 
+use Grav\Common\Grav;
 use Grav\Common\Processors\ProcessorBase;
 use Grav\Framework\Route\Route;
+use Grav\Plugin\Admin\Routers\LoginRouter;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -12,6 +14,16 @@ class Router extends ProcessorBase
 {
     public $id = 'admin_router';
     public $title = 'Admin Panel';
+
+    /** @var Admin */
+    protected $admin;
+
+    public function __construct(Grav $container, Admin $admin)
+    {
+        parent::__construct($container);
+
+        $this->admin = $admin;
+    }
 
     /**
      * Handle routing to the dashboard, group and build objects.
@@ -25,31 +37,32 @@ class Router extends ProcessorBase
         $this->startTimer();
 
         $context = $request->getAttributes();
+        $query = $request->getQueryParams();
 
         /** @var Route $route */
         $route = $context['route'];
         $normalized = mb_strtolower(trim($route->getRoute(), '/'));
         $parts = explode('/', $normalized);
-        array_shift($parts);
-        $key = array_shift($parts);
+        array_shift($parts); // Admin path
+        $routeStr = implode('/', $parts);
+        $view = array_shift($parts);
         $path = implode('/', $parts);
+        $task = $this->container['task'] ?? $query['task'] ?? null;
+        $action = $this->container['action'] ?? $query['action'] ?? null;
 
-        $request = $request->withAttribute('admin', ['path' => $path, 'parts' => $parts]);
+        $params = ['view' => $view, 'route' => $routeStr, 'path' => $path, 'parts' => $parts, 'task' => $task, 'action' => $action];
+        $request = $request->withAttribute('admin', $params);
 
-        $response = null;
-        /*
-        if ($key === '__TODO__') {
-
-            $controller = new TodoController();
-
-            $response = $controller->handle($request);
+        // Run login controller if user isn't fully logged in or asks to logout.
+        $user = $this->admin->user;
+        if (!$user->authorized || !$user->authorize('admin.login')) {
+            $params = (new LoginRouter())->matchServerRequest($request);
+            $request = $request->withAttribute('admin', $params + $request->getAttribute('admin'));
         }
-        */
 
-        if (!$response) {
-            // Fallback to the old admin behavior.
-            $response = $handler->handle($request);
-        }
+        $this->admin->request = $request;
+
+        $response = $handler->handle($request);
 
         $this->stopTimer();
 
