@@ -100,11 +100,11 @@ class AdminController
     protected $id;
     /** @var bool */
     protected $active;
-    /** @var FlexObjectInterface */
+    /** @var FlexObjectInterface|false|null */
     protected $object;
-    /** @var FlexCollectionInterface */
+    /** @var FlexCollectionInterface|null */
     protected $collection;
-    /** @var FlexDirectoryInterface */
+    /** @var FlexDirectoryInterface|null */
     protected $directory;
 
     /** @var string */
@@ -344,7 +344,6 @@ class AdminController
 
                 $grav = Grav::instance();
                 $grav->fireEvent('onFlexAfterDelete', new Event(['type' => 'flex', 'object' => $object]));
-                $grav->fireEvent('gitsync');
             }
         } catch (RuntimeException $e) {
             $this->admin->setMessage('Delete Failed: ' . $e->getMessage(), 'error');
@@ -409,8 +408,13 @@ class AdminController
 
         /** @var UniformResourceLocator $locator */
         $locator = $this->grav['locator'];
+        if ($locator->isStream($new_path)) {
+            $new_path = $locator->findResource($new_path, true, true);
+        } else {
+            $new_path = GRAV_ROOT . '/' . $new_path;
+        }
 
-        Folder::create($locator->findResource($new_path, true, true));
+        Folder::create($new_path);
         Cache::clearCache('invalidate');
 
         $this->grav->fireEvent('onAdminAfterSaveAs', new Event(['path' => $new_path]));
@@ -588,7 +592,7 @@ class AdminController
 
         // Base64 decode the route
         $data['route'] = isset($data['route']) ? base64_decode($data['route']) : null;
-        $data['filters'] = json_decode($options['filters'] ?? '{}', true) + ['type' => ['root', 'dir']];
+        $data['filters'] = json_decode($options['filters'] ?? '{}', true, 512, JSON_THROW_ON_ERROR) + ['type' => ['root', 'dir']];
 
         $initial = $data['initial'] ?? null;
         if ($initial) {
@@ -734,11 +738,18 @@ class AdminController
             /** @var FlexForm $form */
             $form = $this->getForm($object);
 
-            $callable = static function (array $data, array $files, FlexObject $object) use ($form) {
+            $callable = function (array $data, array $files, FlexObject $object) use ($form) {
+                if (method_exists($object, 'storeOriginal')) {
+                    $object->storeOriginal();
+                }
                 $object->update($data, $files);
 
                 // Support for expert mode.
                 if (str_ends_with($form->getId(), '-raw') && isset($data['frontmatter']) && is_callable([$object, 'frontmatter'])) {
+                    if (!$this->user->authorize('admin.super')) {
+                        throw new RuntimeException($this->admin::translate('PLUGIN_ADMIN.INSUFFICIENT_PERMISSIONS_FOR_TASK') . ' save raw.',
+                        403);
+                    }
                     $object->frontmatter($data['frontmatter']);
                     unset($data['frontmatter']);
                 }
@@ -830,7 +841,6 @@ class AdminController
 
             $grav = Grav::instance();
             $grav->fireEvent('onFlexAfterSave', new Event(['type' => 'flex', 'object' => $object]));
-            $grav->fireEvent('gitsync');
         } catch (RuntimeException $e) {
             $this->admin->setMessage('Save Failed: ' . $e->getMessage(), 'error');
 
@@ -912,7 +922,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('action', 'media.list');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             die($e->getMessage());
         }
@@ -928,7 +938,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('task', 'media.upload');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             die($e->getMessage());
         }
@@ -944,7 +954,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('task', 'media.delete');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             die($e->getMessage());
         }
@@ -968,7 +978,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('task', 'media.copy');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             die($e->getMessage());
         }
@@ -984,7 +994,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('task', 'media.remove');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             die($e->getMessage());
         }
@@ -1019,7 +1029,7 @@ class AdminController
         try {
             $response = $this->forwardMediaTask('action', 'media.picker');
 
-            $this->admin->json_response = json_decode($response->getBody(), false);
+            $this->admin->json_response = json_decode($response->getBody(), false, 512, JSON_THROW_ON_ERROR);
         } catch (Exception $e) {
             $this->admin->json_response = ['success' => false, 'error' => $e->getMessage()];
         }
@@ -1405,14 +1415,14 @@ class AdminController
                         $object->language($language);
                     }
                 }
-            }
 
-            if (is_callable([$object, 'refresh'])) {
-                $object->refresh();
-            }
+                if (is_callable([$object, 'refresh'])) {
+                    $object->refresh();
+                }
 
-            // Get updated object via form.
-            $this->object = $object->getForm()->getObject();
+                // Get updated object via form.
+                $this->object = $object ? $object->getForm()->getObject() : false;
+            }
         }
 
         return $this->object ?: null;
@@ -1695,7 +1705,7 @@ class AdminController
             if (is_array($value)) {
                 $value = $this->jsonDecode($value);
             } else {
-                $value = json_decode($value, true);
+                $value = json_decode($value, true, 512, JSON_THROW_ON_ERROR);
             }
         }
 
